Replace local FilterLP by tiny_dsp version

+++ software/o_c_REV/tiny_dsp.h
@@ -24,7 +24,7 @@
 namespace TDSP {
 
 // Constant multiplier for precision divisions
-const int32_t M = 2048;
+const int16_t M = 2048;
 
 // CF = 1/(2*pi*dt) for cHz
 // sample rate dt = 60 u
@@ -53,7 +53,7 @@ public:
 private:
     int32_t y;
     int32_t alpha;
-    const int32_t M = TDSP::M;
+    const int16_t M = TDSP::M;
     const int32_t CF = TDSP::CF;
 };
